chore: move sort logics to utils
